Fix startpoint sampling for PEtab-derived problems with fixed parameters (#1169)
Startpoint sampling for `PetabImporter`-derived problems didn't work correctly in case any parameters were fixed in addition to those marked `estimate=0` in the underlying PEtab problem.

Fixing any parameters after the construction of the `pypesto.Problem` and the corresponding startpoint method would lead to errors during startpoint sampling because the list of fixed parameters was never updated.

In order to fix that, we need to have the current `pypesto.Problem` available for startpoint sampling to get access to the currently fixed parameters. Accessing `pypesto.Problem` is not compatible with the current `FunctionStartpoints`. Therefore, we derive a new `PetabStartpoints` class from `CheckedStartpoints.sample` that will allow forwarding/accessing the `Problem`.

Co-authored-by: Fabian Fröhlich <fabian@schaluck.com>