Fix waterfall start indices (#1200)
Otherwise, when there is more than one optimization result, one might get different `start_indices` depending on result order. E.g., if  a result with fewer multistarts precedes one with more multistarts, the resulting start indices will be limited to the indices of the first  result. If the order of the results is opposite, the result with fewer multistarts won't limit the number of starts that will be plotted for the other result.