Feature spline regularization (#1222)
* Initial, missing regularization grad

Initial commit with spline regularization.
The implementation of the gradient wrt theta is not completed (swapping to develop for some debugging of censored).

Additionally, small fixes of censored gradient (sigma related, not important usually).

* Complete gradient

Implemented the complete gradient, including the regularization term for the ds_dtheta

* Fix par plot, add reg to spline plot

FIxed the parameter plot issue (will be in a separate PR).
Added the linear regularization to the spline visualization.

* Some random changes TODO

Changes that I might have to revert for the Pull request. The ensemble one might've already been pushed by Polina. The solver one completely circumvents dsdtheta gradient calculation due to some "array cannot be inf or NaN" error which pops sometimes, but it's mostly always 0 so removing it makes it easier to deal with for now. This is a big TODO, do not forget.

* Fix par_sim_idx in spline solver

Fixed the par_sim_idx in the spline solver calculate gradient function. For all conditions, the rdata form is not managed by par_sim_idx, but with par_edata_indices. This is done such that if some parameters are not used in the condition (observable or noise parameters for instance) then the sensitivities wrt. them don't have to be calculated

* Fix par_sim_index, pass edata_indices

Added the passing of the edata indices from the calculator to the solver gradient calculation.
Additionally, fixed the inner calculator collector quantitative calculation with nansum instead of a regular one (sometimes it gave only nan values because of this...)

* Remove ds_dtheta_term calculation

Since I've proven that the ds_dtheta gradient contribution is always 0, we can remove all of the code that was calculating it

* Remove redundancy

* Small cleanup

* Remove a mistake

* Add regularization test

* Quality test fix

* Spline tests fix + obj fun fix

* Notebook update

- Updated the notebook for non-linear semi-quantitative data
- fixed the max_sensi_order problem
- changed default min_diff_factor
- fix quality

* Daniel review changes

* Dilan&Fabian review changes

* Improve test coverage

Added tests for untested lines in calculators

---------

Co-authored-by: Stephan Grein <stephan.grein@uni-bonn.de>