Hierarchical optimization: Support for box constraints on offset and scaling parameters (#1238)
* Initial working

Implemented scaling and offset bounds for the analytical and numerical solver of the hierarchical problem.

* Fix hierarchical test

* Add test_constrained_inner_solver

* Add tests for non coupled parameters

* intentional failure

* intentional failure

* intentional failure

* Test fix

It seems L-BFGS-B would not converge to optimum in some cases if it was provided with the big bounds [-1e20, 1e20]. It seems like it is a scipy issue? Not sure.
In any case, the numerical optimizer is better if the optimization is not provided with these dummy bounds. But in this case, we need to sample the bounds outside of the pypesto minimize function and provide it as x_guesses.

* Add notebook changes

* Update C.py

* Revert "Update C.py"

This reverts commit f099a2b1e053502da3267e4cf8f152e5ac334a0f.

* Fix sampling + test

* Cleanup

* Some updates

* Some more small updates

* Add inner bounds to parameter plot

* Fix import issues

* Fix hier test

* Fix parameter plot test

* Remove figures

* DIlan&Daniel review changes

* Update parameters.py

* Update solver.py