Remove CESSOptimizer
SacessOptimizer is superior to CESSOptimizer and I wouldn't recommend using it.
I think it's best to just remove it. No value, just extra maintenance effort.
I don't think anybody was using it yet anyways, so this shouldn't actually be a
breaking change for anybody. If not, please let me know.